cryptography: add PGP signing/verification
Adds the PGP signing scheme. Same logic
as the "full" mode, the user needs to provide the desired
PGP armored private key, which will create a base64
encoded PGP signature embed in the .pack.

In order to verify it, the public GPG key must be
present locally and provided.
