+++ cmd/commands/signature.go
@@ -14,9 +14,6 @@ var signatureCreateflags struct {
 	// certOnly skips private key usage
 	certOnly bool
 
-	//TODO
-	//pgp bool
-
 	// certPath points to the signer's certificate
 	certPath string
 
@@ -25,6 +22,9 @@ var signatureCreateflags struct {
 
 	// outputDir saves the signed pack to a specific path
 	outputDir string
+
+	// pgp mode embeds a PGP signature instead
+	pgp bool
 
 	// skipCertValidation skips sanity/safety checks on the provided certificate
 	skipCertValidation bool
@@ -38,28 +38,26 @@ var signatureVerifyflags struct {
 	export bool
 
 	// pgpKey loads a PGP public key to verify against the signature
-	// pgpKey string
+	pgpKey string
 
 	// skipCertValidation skips sanity/safety checks on the provided certificate
 	skipCertValidation bool
 
 	// skipInfo skips displaying certificate info
 	skipInfo bool
-
-	// TODO:
-	//skipAddKeychain bool
 }
 
 func init() {
-	SignatureCreateCmd.Flags().BoolVar(&signatureCreateflags.certOnly, "cert-only", false, "certificate-only signature")
-	SignatureCreateCmd.Flags().StringVarP(&signatureCreateflags.certPath, "certificate", "c", "", "path for the signer's certificate")
-	_ = SignatureCreateCmd.MarkFlagRequired("certificate")
-	SignatureCreateCmd.Flags().StringVarP(&signatureCreateflags.keyPath, "key", "k", "", "path for the signer's private key")
+	SignatureCreateCmd.Flags().BoolVar(&signatureCreateflags.certOnly, "cert-only", false, "certificate-only signature mode")
+	SignatureCreateCmd.Flags().StringVarP(&signatureCreateflags.certPath, "certificate", "c", "", "path of the signer's certificate")
+	SignatureCreateCmd.Flags().StringVarP(&signatureCreateflags.keyPath, "key", "k", "", "path of the signer's private key")
 	SignatureCreateCmd.Flags().StringVarP(&signatureCreateflags.outputDir, "output-dir", "o", "", "save the signed pack to a specific path")
+	SignatureCreateCmd.Flags().BoolVar(&signatureCreateflags.pgp, "pgp", false, "PGP signature mode")
 	SignatureCreateCmd.Flags().BoolVar(&signatureCreateflags.skipCertValidation, "skip-validation", false, "do not validate certificate")
 	SignatureCreateCmd.Flags().BoolVar(&signatureCreateflags.skipInfo, "skip-info", false, "do not display certificate information")
 
 	SignatureVerifyCmd.Flags().BoolVarP(&signatureVerifyflags.export, "export", "e", false, "only export embed certificate")
+	SignatureVerifyCmd.Flags().StringVarP(&signatureVerifyflags.pgpKey, "pub-key", "k", "", "path of the PGP public key")
 	SignatureVerifyCmd.Flags().BoolVar(&signatureVerifyflags.skipCertValidation, "skip-validation", false, "do not validate certificate")
 	SignatureVerifyCmd.Flags().BoolVar(&signatureVerifyflags.skipInfo, "skip-info", false, "do not display certificate information")
 
@@ -75,25 +73,54 @@ func init() {
 }
 
 var SignatureCreateCmd = &cobra.Command{
-	// TODO: refactor to generic SignatureCreateCmd (default full, --cert-only or --pgp to specify)
 	Use:   "signature-create [<local .path pack>]",
-	Short: "Digitally signs a pack with a X509 certificate",
+	Short: "Digitally signs a pack with a X509 certificate or PGP key",
 	Long: `
-Signs a pack using a x509 certificate and its private key.
+Signs a pack using X509 Public Key Infrastructure or PGP signatures.
+
+Three modes are available. "full" is the default, and takes a X509 public key
+certificate and its private key (currently only RSA supported), which is used
+to sign the hashed (SHA256) contents of a pack.
+If "--cert-only" is specified, only a X509 certificate will be embed in the pack. This
+offers a lesser degree of security guarantees.
+Both these options perform some basic validations on the X509 certificate, which can
+be skipped.
+
+If "--pgp" is specified, the user must provide a PGP private key (Curve25519 or RSA 2048,
+3072 and 4096 bits are supported).
+
+The signature follows a simple scheme which includes the cpackget version used to sign,
+the mode, and the outputs, base64 encoded - saved to the pack's Zip comment field.
+These can be viewed with any text/hex editor or dedicated zip tools like "zipinfo".
 
 The referenced pack must be in its original/compressed form (.pack), and be present locally:
 
-  $ cpackget signature-create Vendor.Pack.1.2.3.pack`,
+  $ cpackget signature-create Vendor.Pack.1.2.3.pack -k private.key -c certificate.pem`,
 	Args: cobra.ExactArgs(1),
 	RunE: func(cmd *cobra.Command, args []string) error {
+		// TODO: skipCertinfo and validation
 		if signatureCreateflags.keyPath == "" {
-			if !signatureCreateflags.certOnly {
-				log.Error("specify private key file with the --key/-k flag")
+			if !signatureCreateflags.certOnly || signatureCreateflags.pgp {
+				log.Error("specify private key file with the -k/-key flag")
 				return errs.ErrIncorrectCmdArgs
 			}
 		} else {
 			if signatureCreateflags.certOnly {
 				log.Error("-k/--key should not be provided in certificate-only mode")
+				return errs.ErrIncorrectCmdArgs
+			}
+		}
+		if signatureCreateflags.pgp {
+			if signatureCreateflags.certPath != "" {
+				log.Error("PGP signature scheme does not need a x509 certificate")
+				return errs.ErrIncorrectCmdArgs
+			}
+			if signatureCreateflags.skipCertValidation {
+				log.Error("PGP signature scheme does not validate certificates (--skip-validation)")
+				return errs.ErrIncorrectCmdArgs
+			}
+			if signatureCreateflags.skipInfo {
+				log.Error("PGP signature scheme does not display certificate info (--skip-info)")
 				return errs.ErrIncorrectCmdArgs
 			}
 		}
@@ -108,6 +135,11 @@ var SignatureVerifyCmd = &cobra.Command{
 Verifies the integrity and authenticity of a pack signed
 with the "signature-create" command.
 
+For more information on the signatures, use "cpackget help signature-create".
+
+If attempting to verify a PGP signed pack, use the -k/--pub-key flag to specify
+the publisher's public GPG key.
+
 The referenced pack must be in its original/compressed form (.pack), and be present locally:
 
   $ cpackget signature-verify Vendor.Pack.1.2.3.pack.signed`,
@@ -117,6 +149,16 @@ The referenced pack must be in its original/compressed form (.pack), and be pres
 			log.Error("-e/--export does not need any other flags")
 			return errs.ErrIncorrectCmdArgs
 		}
-		return cryptography.VerifyPackSignature(args[0], Version, signatureVerifyflags.export, signatureVerifyflags.skipCertValidation, signatureVerifyflags.skipInfo)
+		if signatureVerifyflags.pgpKey != "" {
+			if signatureVerifyflags.export {
+				log.Error("can't export non X509 (full, cert-only) signature scheme")
+				return errs.ErrIncorrectCmdArgs
+			}
+			if signatureVerifyflags.skipCertValidation || signatureVerifyflags.skipInfo {
+				log.Error("PGP verification does not need any flags other than -k/-pub-key")
+				return errs.ErrIncorrectCmdArgs
+			}
+		}
+		return cryptography.VerifyPackSignature(args[0], signatureVerifyflags.pgpKey, Version, signatureVerifyflags.export, signatureVerifyflags.skipCertValidation, signatureVerifyflags.skipInfo)
 	},
 }
